Make RawTime type link time discoverable
