+++ Os/CMakeLists.txt
@@ -143,7 +143,24 @@ add_named_os_module(CountingSemaphore)
 add_named_os_module(Queue)
 add_named_os_module(Cpu)
 add_named_os_module(Memory)
-add_named_os_module(RawTime Fw_Buffer)
+# RawTime: RawTime.hpp is a thin public alias header (no .cpp); the interface and
+# link-time delegate live in DelegateRawTime.{hpp,cpp}. Registered explicitly
+# because add_named_os_module assumes a matching <name>.cpp per header.
+register_fprime_module(
+      "Os_RawTime"
+    REQUIRES_IMPLEMENTATIONS
+      "Os_RawTime"
+    SOURCES
+      "${CMAKE_CURRENT_LIST_DIR}/DelegateRawTime.cpp"
+    HEADERS
+      "${CMAKE_CURRENT_LIST_DIR}/RawTime.hpp"
+      "${CMAKE_CURRENT_LIST_DIR}/DelegateRawTime.hpp"
+    DEPENDS
+      Fw_Time
+      Fw_Types
+      Fw_Buffer
+)
+fprime_target_dependencies(Os PUBLIC "Os_RawTime")
 
 ### UTS ### Note: 3 separate UTs registered here.
 register_fprime_ut(
